Added Growth length to gdm. Fix to rebuild of KD-tree when the tree is empty.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/Plugins/AngioFE2@9305 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ AngioFE2/FEAngioMaterial.cpp
@@ -900,7 +900,7 @@ mat3ds FEAngioMaterial::AngioStress(FEAngioMaterialPoint& angioPt)
 			temp.push_back(stemp);
 			std::pair<size_t, std::vector<SPROUT> *> dim = std::pair<size_t, std::vector<SPROUT> * >(0, &temp);
 			std::vector<std::pair<size_t, std::vector<SPROUT> *>> nst = sprouts.within(dim, m_cultureParams.stress_radius * m_cultureParams.stress_radius);
-			for (int i = 0; i<nst.size(); ++i)
+			for (size_t i = 0; i<nst.size(); ++i)
 			{
 				SPROUT& sp = m_spr[nst[i].first];
